refactor: JSONParser, MemoDecodeModel 구조체 리팩토링
- JSONParser의 파라미터 타입 수정 (T -> T.Type)
